+++ Civi/ActionProvider/Action/Activity/CreateActivity.php
@@ -44,6 +44,7 @@ class CreateActivity extends AbstractAction {
       new Specification('activity_date', 'Timestamp', E::ts('Activity Date'), TRUE),
       new Specification('id', 'Integer', E::ts('Activity ID'), false),
       new Specification('campaign_id', 'Integer', E::ts('Campaign'), false),
+      new Specification('details', 'Text', E::ts('Details'), false),
     ]);
 
     $customGroups = civicrm_api3('CustomGroup', 'get', [
@@ -110,6 +111,9 @@ class CreateActivity extends AbstractAction {
     $activityParams['activity_date_time'] = $parameters->getParameter('activity_date');
     if ($parameters->doesParameterExists('campaign_id')) {
       $activityParams['campaign_id'] = $parameters->getParameter('campaign_id');
+    }
+    if ($parameters->doesParameterExists('details')) {
+      $activityParams['details'] = $parameters->getParameter('details');
     }
 
     foreach($this->getParameterSpecification() as $spec) {